Introduce an interface DirectoryFiller and move implementation into a new class to make implementations of readdir() in FuseFilesystem unit-testable
